refactor: keep React root tracking out of the public API
The registerRenderedRoot/takeRenderedRoots helpers (and the react-dom/client
Root type in their signatures) are internal coordination for
AdaptiveCardRenderer, but they were re-exported through utils/index.ts ->
src/index.ts, making a private rendering detail part of the package's public,
semver-sensitive API.

Move them into a private utils/rendered-roots.ts that is not re-exported from
the utils barrel, and import them directly where needed. createInputContainer
stays in renderer-utils.ts.
